extracted multiple new headerfiles from roadmanager to make it easier to work with
