--- EnvironmentSimulator/Modules/RoadManager/RoadManager.hpp
@@ -1,1727 +1,0 @@
-﻿/*
- * esmini - Environment Simulator Minimalistic
- * https://github.com/esmini/esmini
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * Copyright (c) partners of Simulation Scenarios
- * https://sites.google.com/view/simulationscenarios
- */
-
-#ifndef OPENDRIVE_HH_
-#define OPENDRIVE_HH_
-
-#include <cmath>
-#include <experimental/optional>
-#include <list>
-#include <map>
-#include <string>
-#include <vector>
-#include "CommonMini.hpp"
-#include "pugixml.hpp"
-
-#define PARAMPOLY3_STEPS 100
-
-namespace roadmanager {
-int GetNewGlobalLaneId();
-int GetNewGlobalLaneBoundaryId();
-
-class UserData {
-   public:
-	UserData(std::string code, std::string value, pugi::xml_node userDataNode) : code_(code), value_(value) {
-		origin_node_ = pugi::xml_document();
-		origin_node_.append_copy(userDataNode);
-	}
-
-	UserData(pugi::xml_node userDataNode) {
-		code_ = userDataNode.attribute("code").as_string();
-		value_ = userDataNode.attribute("value").as_string();
-		origin_node_ = pugi::xml_document();
-		origin_node_.append_copy(userDataNode);
-	}
-
-	std::string GetCode() const { return code_; }
-	std::string GetValue() const { return value_; }
-
-	void Save(pugi::xml_node& parent);
-
-   private:
-	std::string code_;
-	std::string value_;
-	pugi::xml_document origin_node_;
-};
-
-typedef struct {
-	double s;
-	double x;
-	double y;
-	double z;
-	double h;
-} PointStruct;
-class Elevation {
-   public:
-	Elevation() : s_(0.0), length_(0.0) {}
-	Elevation(double s, double a, double b, double c, double d) : s_(s), length_(0) {
-		poly3_.Set(a, b, c, d);
-	}
-	~Elevation(){};
-
-	double GetS() { return s_; }
-	void SetLength(double length) { length_ = length; }
-	double GetLength() { return length_; }
-	void Print();
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node&, const std::string);
-	Polynomial poly3_;
-
-   private:
-	double s_;
-	double length_;
-	std::vector<UserData*> user_data_;
-};
-
-typedef enum { NONE = 0, SUCCESSOR = 1, PREDECESSOR = -1 } LinkType;
-
-class LaneLink {
-   public:
-	LaneLink(LinkType type, int id) : type_(type), id_(id) {}
-
-	LinkType GetType() { return type_; }
-	int GetId() { return id_; }
-	void Print();
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node&);
-
-   private:
-	LinkType type_;
-	int id_;
-	std::vector<UserData*> user_data_;
-};
-
-class LaneWidth {
-   public:
-	LaneWidth(double s_offset, double a, double b, double c, double d) : s_offset_(s_offset) {
-		poly3_.Set(a, b, c, d);
-	}
-
-	double GetSOffset() { return s_offset_; }
-	void Print();
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node&);
-
-	Polynomial poly3_;
-
-   private:
-	double s_offset_;
-	std::vector<UserData*> user_data_;
-};
-
-struct RoadMarkInfo {
-	int roadmark_idx_;
-	int roadmarkline_idx_;
-};
-
-enum class RoadMarkColor {
-	UNDEFINED,
-	STANDARD_COLOR,	 // equivalent to white
-	BLUE,
-	GREEN,
-	RED,
-	WHITE,
-	YELLOW
-};
-
-class LaneRoadMarkTypeLine {
-   public:
-	enum RoadMarkTypeLineRule { NO_PASSING, CAUTION, NONE };
-
-	LaneRoadMarkTypeLine(double length,
-						 double space,
-						 double t_offset,
-						 double s_offset,
-						 RoadMarkTypeLineRule rule,
-						 double width,
-						 RoadMarkColor color = RoadMarkColor::UNDEFINED)
-		: length_(length),
-		  space_(space),
-		  t_offset_(t_offset),
-		  s_offset_(s_offset),
-		  rule_(rule),
-		  width_(width),
-		  color_(color) {}
-	~LaneRoadMarkTypeLine(){};
-	double GetSOffset() { return s_offset_; }
-	double GetTOffset() { return t_offset_; }
-	double GetLength() { return length_; }
-	double GetSpace() { return space_; }
-	double GetWidth() { return width_; }
-	OSIPoints* GetOSIPoints() { return &osi_points_; }
-	OSIPoints osi_points_;
-	void SetGlobalId();
-	int GetGlobalId() { return global_id_; }
-	RoadMarkColor GetColor() { return color_; }
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node&);
-
-   private:
-	double length_;
-	double space_;
-	double t_offset_;
-	double s_offset_;
-	RoadMarkTypeLineRule rule_;
-	double width_;
-	int global_id_;		   // Unique ID for OSI
-	RoadMarkColor color_;  // if set, supersedes setting in <RoadMark>
-	std::vector<UserData*> user_data_;
-};
-
-class LaneRoadMarkType {
-   public:
-	LaneRoadMarkType(std::string name, double width) : name_(name), width_(width) {}
-
-	void AddLine(LaneRoadMarkTypeLine* lane_roadMarkTypeLine);
-	std::string GetName() { return name_; }
-	double GetWidth() { return width_; }
-	LaneRoadMarkTypeLine* GetLaneRoadMarkTypeLineByIdx(int idx);
-	int GetNumberOfRoadMarkTypeLines() { return (int)lane_roadMarkTypeLine_.size(); }
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node&);
-
-   private:
-	std::string name_;
-	double width_;
-	std::vector<LaneRoadMarkTypeLine*> lane_roadMarkTypeLine_;
-	std::vector<UserData*> user_data_;
-};
-
-class LaneRoadMark {
-   public:
-	enum RoadMarkType {
-		NONE_TYPE = 1,
-		SOLID = 2,
-		BROKEN = 3,
-		SOLID_SOLID = 4,
-		SOLID_BROKEN = 5,
-		BROKEN_SOLID = 6,
-		BROKEN_BROKEN = 7,
-		BOTTS_DOTS = 8,
-		GRASS = 9,
-		CURB = 10
-	};
-
-	enum RoadMarkWeight { STANDARD, BOLD };
-
-	enum RoadMarkMaterial {
-		STANDARD_MATERIAL  // only "standard" is available for now
-	};
-
-	enum RoadMarkLaneChange { INCREASE, DECREASE, BOTH, NONE_LANECHANGE };
-
-	LaneRoadMark(double s_offset,
-				 RoadMarkType type,
-				 RoadMarkWeight weight,
-				 RoadMarkColor color,
-				 RoadMarkMaterial material,
-				 RoadMarkLaneChange lane_change,
-				 double width,
-				 double height)
-		: s_offset_(s_offset),
-		  type_(type),
-		  weight_(weight),
-		  color_(color),
-		  material_(material),
-		  lane_change_(lane_change),
-		  width_(width),
-		  height_(height) {}
-
-	void AddType(LaneRoadMarkType* lane_roadMarkType) { lane_roadMarkType_.push_back(lane_roadMarkType); }
-
-	double GetSOffset() { return s_offset_; }
-	double GetWidth() { return width_; }
-	double GetHeight() { return height_; }
-	RoadMarkType GetType() { return type_; }
-	RoadMarkWeight GetWeight() { return weight_; }
-	RoadMarkColor GetColor() { return color_; }
-	RoadMarkMaterial GetMaterial() { return material_; }
-	RoadMarkLaneChange GetLaneChange() { return lane_change_; }
-
-	int GetNumberOfRoadMarkTypes() { return (int)lane_roadMarkType_.size(); }
-	LaneRoadMarkType* GetLaneRoadMarkTypeByIdx(int idx);
-
-	static RoadMarkColor ParseColor(pugi::xml_node node);
-
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node&);
-
-   private:
-	double s_offset_;
-	RoadMarkType type_;
-	RoadMarkWeight weight_;
-	RoadMarkColor color_;
-	RoadMarkMaterial material_;
-	RoadMarkLaneChange lane_change_;
-	double width_;
-	double height_;
-	std::vector<LaneRoadMarkType*> lane_roadMarkType_;
-	std::vector<UserData*> user_data_;
-};
-
-class LaneMaterial {
-   public:
-	LaneMaterial(double sOffset, std::string surface, double friction, double roughness)
-		: s_offset_(sOffset), friction_(friction), roughness_(roughness), surface_(surface){};
-
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node& lane);
-
-   private:
-	double s_offset_;
-	std::string surface_;
-	double friction_;
-	double roughness_;
-	std::vector<UserData*> user_data_;
-};
-
-class LaneSpeed {
-   public:
-	LaneSpeed(double sOffset, double max, std::string unit) : s_offset_(sOffset), max_(max), unit_(unit){};
-
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node& lane);
-
-   private:
-	double s_offset_;
-	double max_;
-	std::string unit_;
-	std::vector<UserData*> user_data_;
-};
-
-class LaneOffset {
-   public:
-	LaneOffset() : s_(0.0), length_(0.0) {}
-	LaneOffset(double s, double a, double b, double c, double d) : s_(s), length_(0.0) {
-		polynomial_.Set(a, b, c, d);
-	}
-	~LaneOffset() {}
-
-	void Set(double s, double a, double b, double c, double d) {
-		s_ = s;
-		polynomial_.Set(a, b, c, d);
-	}
-	void SetLength(double length) { length_ = length; }
-	double GetS() { return s_; }
-	Polynomial GetPolynomial() { return polynomial_; }
-	double GetLength() { return length_; }
-	double GetLaneOffset(double s);
-	double GetLaneOffsetPrim(double s);
-	void Print();
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node&);
-
-   private:
-	Polynomial polynomial_;
-	double s_;
-	double length_;
-	std::vector<UserData*> user_data_;
-};
-
-class Lane {
-   public:
-	enum LanePosition { LANE_POS_CENTER, LANE_POS_LEFT, LANE_POS_RIGHT };
-
-	typedef enum {
-		LANE_TYPE_NONE = (1 << 0),
-		LANE_TYPE_DRIVING = (1 << 1),
-		LANE_TYPE_STOP = (1 << 2),
-		LANE_TYPE_SHOULDER = (1 << 3),
-		LANE_TYPE_BIKING = (1 << 4),
-		LANE_TYPE_SIDEWALK = (1 << 5),
-		LANE_TYPE_BORDER = (1 << 6),
-		LANE_TYPE_RESTRICTED = (1 << 7),
-		LANE_TYPE_PARKING = (1 << 8),
-		LANE_TYPE_BIDIRECTIONAL = (1 << 9),
-		LANE_TYPE_MEDIAN = (1 << 10),
-		LANE_TYPE_SPECIAL1 = (1 << 11),
-		LANE_TYPE_SPECIAL2 = (1 << 12),
-		LANE_TYPE_SPECIAL3 = (1 << 13),
-		LANE_TYPE_ROADWORKS = (1 << 14),
-		LANE_TYPE_TRAM = (1 << 15),
-		LANE_TYPE_RAIL = (1 << 16),
-		LANE_TYPE_ENTRY = (1 << 17),
-		LANE_TYPE_EXIT = (1 << 18),
-		LANE_TYPE_OFF_RAMP = (1 << 19),
-		LANE_TYPE_ON_RAMP = (1 << 20),
-		LANE_TYPE_ANY_DRIVING = LANE_TYPE_DRIVING | LANE_TYPE_ENTRY | LANE_TYPE_EXIT | LANE_TYPE_OFF_RAMP
-								| LANE_TYPE_ON_RAMP | LANE_TYPE_PARKING,
-		LANE_TYPE_ANY_ROAD = LANE_TYPE_ANY_DRIVING | LANE_TYPE_RESTRICTED | LANE_TYPE_STOP,
-		LANE_TYPE_ANY = (0xFFFFFFFF)
-	} LaneType;
-
-	// Construct & Destruct
-	Lane()
-		: id_(0),
-		  type_(LaneType::LANE_TYPE_NONE),
-		  level_(-1),
-		  offset_from_ref_(0.0),
-		  global_id_(0),
-		  lane_boundary_(0) {}
-	Lane(int id, Lane::LaneType type)
-		: id_(id), type_(type), level_(-1), offset_from_ref_(0), global_id_(0), lane_boundary_(0) {}
-	~Lane() {}
-
-	// Base Get Functions
-	int GetId() { return id_; }
-	double GetOffsetFromRef() { return offset_from_ref_; }
-	LaneType GetLaneType() { return type_; }
-	int GetGlobalId() { return global_id_; }
-
-	// Add Functions
-	void AddLink(LaneLink* lane_link) { link_.push_back(lane_link); }
-	void AddLaneWidth(LaneWidth* lane_width) { lane_width_.push_back(lane_width); }
-	void AddLaneRoadMark(LaneRoadMark* lane_roadMark) { lane_roadMark_.push_back(lane_roadMark); }
-	void AddLaneMaterial(LaneMaterial* lane_material) { lane_material_.push_back(lane_material); }
-	void AddLaneSpeed(LaneSpeed* lane_speed) { lane_speed_.push_back(lane_speed); }
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	// Get Functions
-	int GetNumberOfRoadMarks() { return (int)lane_roadMark_.size(); }
-	int GetNumberOfLinks() { return (int)link_.size(); }
-	int GetNumberOfLaneWidths() { return (int)lane_width_.size(); }
-
-	LaneLink* GetLink(LinkType type);
-	LaneWidth* GetWidthByIndex(int index);
-	LaneWidth* GetWidthByS(double s);
-	LaneRoadMark* GetLaneRoadMarkByIdx(int idx);
-
-	RoadMarkInfo GetRoadMarkInfoByS(int track_id, int lane_id, double s);
-	OSIPoints* GetOSIPoints() { return &osi_points_; }
-	std::vector<int> GetLineGlobalIds();
-	LaneBoundaryOSI* GetLaneBoundary() { return lane_boundary_; }
-	int GetLaneBoundaryGlobalId();
-
-	// Set Functions
-	void SetGlobalId();
-	void SetLaneBoundary(LaneBoundaryOSI* lane_boundary);
-	void SetOffsetFromRef(double offset) { offset_from_ref_ = offset; }
-	void SetLevel(int level) { level_ = level; }
-
-	// Others
-	bool IsType(Lane::LaneType type);
-	bool IsCenter();
-	bool IsDriving();
-	bool IsOSIIntersection() { return osiintersection_ > 0; }
-	void SetOSIIntersection(int is_osi_intersection) { osiintersection_ = is_osi_intersection; }
-	int GetOSIIntersectionId() { return osiintersection_; }
-	void Print();
-	void Save(pugi::xml_node&);
-	OSIPoints osi_points_;
-
-   private:
-	int id_;			   // center = 0, left > 0, right < 0
-	int global_id_;		   // Unique ID for OSI
-	int osiintersection_;  // flag to see if the lane is part of an osi-lane section or not
-	LaneType type_;
-	int level_;	 // boolean, true = keep lane on level
-	double offset_from_ref_;
-	std::vector<LaneLink*> link_;
-	std::vector<LaneWidth*> lane_width_;
-	std::vector<LaneRoadMark*> lane_roadMark_;
-	std::vector<LaneMaterial*> lane_material_;
-	std::vector<LaneSpeed*> lane_speed_;
-	LaneBoundaryOSI* lane_boundary_;
-	std::vector<UserData*> user_data_;
-};
-
-class LaneSection {
-   public:
-	LaneSection(double s) : s_(s), length_(0) {}
-	void AddLane(Lane* lane);
-	double GetS() { return s_; }
-	Lane* GetLaneByIdx(int idx);
-	Lane* GetLaneById(int id);
-	int GetLaneIdByIdx(int idx);
-	int GetLaneIdxById(int id);
-	bool IsOSILaneById(int id);
-	int GetLaneGlobalIdByIdx(int idx);
-	int GetLaneGlobalIdById(int id);
-	double GetOuterOffset(double s, int lane_id);
-	double GetWidth(double s, int lane_id);
-	int GetClosestLaneIdx(double s,
-						  double t,
-						  double& offset,
-						  bool noZeroWidth,
-						  int laneTypeMask = Lane::LaneType::LANE_TYPE_ANY_DRIVING);
-
-	/**
-	Get lateral position of lane center, from road reference lane (lane id=0)
-	Example: If lane id 1 is 5 m wide and lane id 2 is 4 m wide, then
-			lane 1 center offset is 5/2 = 2.5 and lane 2 center offset is 5 + 4/2 = 7
-	@param s distance along the road segment
-	@param lane_id lane specifier, starting from center -1, -2, ... is on the right side, 1, 2... on the left
-	*/
-	double GetCenterOffset(double s, int lane_id);
-	double GetOuterOffsetHeading(double s, int lane_id);
-	double GetCenterOffsetHeading(double s, int lane_id);
-	double GetLength() { return length_; }
-	int GetNumberOfLanes() { return (int)lane_.size(); }
-	int GetNumberOfDrivingLanes();
-	int GetNumberOfDrivingLanesSide(int side);
-	int GetNUmberOfLanesRight();
-	int GetNUmberOfLanesLeft();
-	void SetLength(double length) { length_ = length; }
-	void SetSingleSide(bool single) { singleSide_ = single; }
-	int GetConnectingLaneId(int incoming_lane_id, LinkType link_type);
-	double GetWidthBetweenLanes(int lane_id1, int lane_id2, double s);
-	double GetOffsetBetweenLanes(int lane_id1, int lane_id2, double s);
-	void Print();
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node&);
-
-   private:
-	double s_;
-	std::experimental::optional<bool> singleSide_;
-	double length_;
-	std::vector<Lane*> lane_;
-	std::vector<UserData*> user_data_;
-};
-
-enum ContactPointType {
-	CONTACT_POINT_UNDEFINED,
-	CONTACT_POINT_START,
-	CONTACT_POINT_END,
-	CONTACT_POINT_JUNCTION,	 // No contact point for element type junction
-};
-
-class RoadLink {
-   public:
-	typedef enum {
-		ELEMENT_TYPE_UNKNOWN,
-		ELEMENT_TYPE_ROAD,
-		ELEMENT_TYPE_JUNCTION,
-	} ElementType;
-
-	RoadLink()
-		: type_(NONE),
-		  element_id_(-1),
-		  element_type_(ELEMENT_TYPE_UNKNOWN),
-		  contact_point_type_(CONTACT_POINT_UNDEFINED) {}
-	RoadLink(LinkType type, ElementType element_type, int element_id, ContactPointType contact_point_type)
-		: type_(type),
-		  element_id_(element_id),
-		  element_type_(element_type),
-		  contact_point_type_(contact_point_type) {}
-	RoadLink(LinkType type, pugi::xml_node node);
-	bool operator==(RoadLink& rhs);
-
-	int GetElementId() { return element_id_; }
-	LinkType GetType() { return type_; }
-	RoadLink::ElementType GetElementType() { return element_type_; }
-	ContactPointType GetContactPointType() { return contact_point_type_; }
-
-	void Print();
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node&);
-
-   private:
-	LinkType type_;
-	int element_id_;
-	ElementType element_type_;
-	ContactPointType contact_point_type_;
-	std::vector<UserData*> user_data_;
-};
-
-struct LaneInfo {
-	int lane_section_idx_;
-	int lane_id_;
-};
-
-typedef struct {
-	int fromLane_;
-	int toLane_;
-	std::vector<UserData*> user_data_;
-	void Save(pugi::xml_node& object);
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-} ValidityRecord;
-
-class RoadObject {
-   public:
-	enum Orientation {
-		POSITIVE,
-		NEGATIVE,
-		NONE,
-	};
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	std::vector<ValidityRecord> validity_;
-	std::vector<UserData*> user_data_;
-};
-
-class ObjectReference : public RoadObject {
-   public:
-	ObjectReference(double s, double t, int id, double zOffset, double validLength, Orientation orientation)
-		: s_(s), t_(t), id_(id), z_offset_(zOffset), valid_length_(validLength), orientation_(orientation) {}
-
-	virtual ~ObjectReference() = default;
-
-	double GetS() const { return s_; }
-	double GetT() const { return t_; }
-	int GetId() const { return id_; }
-	double GetZOffset() const { return z_offset_; }
-	double GetValidLength() const { return valid_length_; }
-	Orientation GetOrientation() const { return orientation_; }
-
-	void Save(pugi::xml_node& objects);
-
-   private:
-	double s_;
-	double t_;
-	int id_;
-	double z_offset_;
-	double valid_length_;
-	Orientation orientation_;
-};
-
-class Bridge : public RoadObject {
-   public:
-	enum Type { CONCRETE, STEEL, BRICK, WOOD, UNKNOWN };
-
-	Bridge(double s, double length, std::string name, int id, Type type)
-		: s_(s), length_(length), name_(name), id_(id), type_(type) {}
-
-	virtual ~Bridge() = default;
-
-	double GetS() const { return s_; }
-	double GetLength() const { return length_; }
-	int GetId() const { return id_; }
-	std::string GetName() const { return name_; }
-	Type GetType() const { return type_; }
-
-	void Save(pugi::xml_node& objects);
-
-   private:
-	double s_;
-	double length_;
-	std::string name_;
-	Type type_;
-	int id_;
-};
-
-class Signal : public RoadObject {
-   public:
-	enum Type : int {
-		TYPE_UNKNOWN = 0,
-		TYPE_OTHER = 1,
-		TYPE_DANGER_SPOT = 2,
-		TYPE_ZEBRA_CROSSING = 87,
-		TYPE_FLIGHT = 110,
-		TYPE_CATTLE = 200,
-		TYPE_HORSE_RIDERS = 197,
-		TYPE_AMPHIBIANS = 188,
-		TYPE_FALLING_ROCKS = 96,
-		TYPE_SNOW_OR_ICE = 94,
-		TYPE_LOOSE_GRAVEL = 97,
-		TYPE_WATERSIDE = 102,
-		TYPE_CLEARANCE = 210,
-		TYPE_MOVABLE_BRIDGE = 101,
-		TYPE_RIGHT_BEFORE_LEFT_NEXT_INTERSECTION = 3,
-		TYPE_TURN_LEFT = 4,
-		TYPE_TURN_RIGHT = 5,
-		TYPE_DOUBLE_TURN_LEFT = 6,
-		TYPE_DOUBLE_TURN_RIGHT = 7,
-		TYPE_HILL_DOWNWARDS = 8,
-		TYPE_HILL_UPWARDS = 9,
-		TYPE_UNEVEN_ROAD = 93,
-		TYPE_ROAD_SLIPPERY_WET_OR_DIRTY = 95,
-		TYPE_SIDE_WINDS = 98,
-		TYPE_ROAD_NARROWING = 10,
-		TYPE_ROAD_NARROWING_RIGHT = 12,
-		TYPE_ROAD_NARROWING_LEFT = 11,
-		TYPE_ROAD_WORKS = 13,
-		TYPE_TRAFFIC_QUEUES = 100,
-		TYPE_TWO_WAY_TRAFFIC = 14,
-		TYPE_ATTENTION_TRAFFIC_LIGHT = 15,
-		TYPE_PEDESTRIANS = 103,
-		TYPE_CHILDREN_CROSSING = 106,
-		TYPE_CYCLE_ROUTE = 107,
-		TYPE_DEER_CROSSING = 109,
-		TYPE_UNGATED_LEVEL_CROSSING = 144,
-		TYPE_LEVEL_CROSSING_MARKER = 112,
-		TYPE_RAILWAY_TRAFFIC_PRIORITY = 135,
-		TYPE_GIVE_WAY = 16,
-		TYPE_STOP = 17,
-		TYPE_PRIORITY_TO_OPPOSITE_DIRECTION = 18,
-		TYPE_PRIORITY_TO_OPPOSITE_DIRECTION_UPSIDE_DOWN = 19,
-		TYPE_PRESCRIBED_LEFT_TURN = 20,
-		TYPE_PRESCRIBED_RIGHT_TURN = 21,
-		TYPE_PRESCRIBED_STRAIGHT = 22,
-		TYPE_PRESCRIBED_RIGHT_WAY = 24,
-		TYPE_PRESCRIBED_LEFT_WAY = 23,
-		TYPE_PRESCRIBED_RIGHT_TURN_AND_STRAIGHT = 26,
-		TYPE_PRESCRIBED_LEFT_TURN_AND_STRAIGHT = 25,
-		TYPE_PRESCRIBED_LEFT_TURN_AND_RIGHT_TURN = 27,
-		TYPE_PRESCRIBED_LEFT_TURN_RIGHT_TURN_AND_STRAIGHT = 28,
-		TYPE_ROUNDABOUT = 29,
-		TYPE_ONEWAY_LEFT = 30,
-		TYPE_ONEWAY_RIGHT = 31,
-		TYPE_PASS_LEFT = 32,
-		TYPE_PASS_RIGHT = 33,
-		TYPE_SIDE_LANE_OPEN_FOR_TRAFFIC = 128,
-		TYPE_SIDE_LANE_CLOSED_FOR_TRAFFIC = 129,
-		TYPE_SIDE_LANE_CLOSING_FOR_TRAFFIC = 130,
-		TYPE_BUS_STOP = 137,
-		TYPE_TAXI_STAND = 138,
-		TYPE_BICYCLES_ONLY = 145,
-		TYPE_HORSE_RIDERS_ONLY = 146,
-		TYPE_PEDESTRIANS_ONLY = 147,
-		TYPE_BICYCLES_PEDESTRIANS_SHARED_ONLY = 148,
-		TYPE_BICYCLES_PEDESTRIANS_SEPARATED_LEFT_ONLY = 149,
-		TYPE_BICYCLES_PEDESTRIANS_SEPARATED_RIGHT_ONLY = 150,
-		TYPE_PEDESTRIAN_ZONE_BEGIN = 151,
-		TYPE_PEDESTRIAN_ZONE_END = 152,
-		TYPE_BICYCLE_ROAD_BEGIN = 153,
-		TYPE_BICYCLE_ROAD_END = 154,
-		TYPE_BUS_LANE = 34,
-		TYPE_BUS_LANE_BEGIN = 35,
-		TYPE_BUS_LANE_END = 36,
-		TYPE_ALL_PROHIBITED = 37,
-		TYPE_MOTORIZED_MULTITRACK_PROHIBITED = 38,
-		TYPE_TRUCKS_PROHIBITED = 39,
-		TYPE_BICYCLES_PROHIBITED = 40,
-		TYPE_MOTORCYCLES_PROHIBITED = 41,
-		TYPE_MOPEDS_PROHIBITED = 155,
-		TYPE_HORSE_RIDERS_PROHIBITED = 156,
-		TYPE_HORSE_CARRIAGES_PROHIBITED = 157,
-		TYPE_CATTLE_PROHIBITED = 158,
-		TYPE_BUSES_PROHIBITED = 159,
-		TYPE_CARS_PROHIBITED = 160,
-		TYPE_CARS_TRAILERS_PROHIBITED = 161,
-		TYPE_TRUCKS_TRAILERS_PROHIBITED = 162,
-		TYPE_TRACTORS_PROHIBITED = 163,
-		TYPE_PEDESTRIANS_PROHIBITED = 42,
-		TYPE_MOTOR_VEHICLES_PROHIBITED = 43,
-		TYPE_HAZARDOUS_GOODS_VEHICLES_PROHIBITED = 164,
-		TYPE_OVER_WEIGHT_VEHICLES_PROHIBITED = 165,
-		TYPE_VEHICLES_AXLE_OVER_WEIGHT_PROHIBITED = 166,
-		TYPE_VEHICLES_EXCESS_WIDTH_PROHIBITED = 167,
-		TYPE_VEHICLES_EXCESS_HEIGHT_PROHIBITED = 168,
-		TYPE_VEHICLES_EXCESS_LENGTH_PROHIBITED = 169,
-		TYPE_DO_NOT_ENTER = 44,
-		TYPE_SNOW_CHAINS_REQUIRED = 170,
-		TYPE_WATER_POLLUTANT_VEHICLES_PROHIBITED = 171,
-		TYPE_ENVIRONMENTAL_ZONE_BEGIN = 45,
-		TYPE_ENVIRONMENTAL_ZONE_END = 46,
-		TYPE_NO_U_TURN_LEFT = 47,
-		TYPE_NO_U_TURN_RIGHT = 48,
-		TYPE_PRESCRIBED_U_TURN_LEFT = 49,
-		TYPE_PRESCRIBED_U_TURN_RIGHT = 50,
-		TYPE_MINIMUM_DISTANCE_FOR_TRUCKS = 51,
-		TYPE_SPEED_LIMIT_BEGIN = 52,
-		TYPE_SPEED_LIMIT_ZONE_BEGIN = 53,
-		TYPE_SPEED_LIMIT_ZONE_END = 54,
-		TYPE_MINIMUM_SPEED_BEGIN = 55,
-		TYPE_OVERTAKING_BAN_BEGIN = 56,
-		TYPE_OVERTAKING_BAN_FOR_TRUCKS_BEGIN = 57,
-		TYPE_SPEED_LIMIT_END = 58,
-		TYPE_MINIMUM_SPEED_END = 59,
-		TYPE_OVERTAKING_BAN_END = 60,
-		TYPE_OVERTAKING_BAN_FOR_TRUCKS_END = 61,
-		TYPE_ALL_RESTRICTIONS_END = 62,
-		TYPE_NO_STOPPING = 63,
-		TYPE_NO_PARKING = 64,
-		TYPE_NO_PARKING_ZONE_BEGIN = 65,
-		TYPE_NO_PARKING_ZONE_END = 66,
-		TYPE_RIGHT_OF_WAY_NEXT_INTERSECTION = 67,
-		TYPE_RIGHT_OF_WAY_BEGIN = 68,
-		TYPE_RIGHT_OF_WAY_END = 69,
-		TYPE_PRIORITY_OVER_OPPOSITE_DIRECTION = 70,
-		TYPE_PRIORITY_OVER_OPPOSITE_DIRECTION_UPSIDE_DOWN = 71,
-		TYPE_TOWN_BEGIN = 72,
-		TYPE_TOWN_END = 73,
-		TYPE_CAR_PARKING = 74,
-		TYPE_CAR_PARKING_ZONE_BEGIN = 75,
-		TYPE_CAR_PARKING_ZONE_END = 76,
-		TYPE_SIDEWALK_HALF_PARKING_LEFT = 172,
-		TYPE_SIDEWALK_HALF_PARKING_RIGHT = 173,
-		TYPE_SIDEWALK_PARKING_LEFT = 174,
-		TYPE_SIDEWALK_PARKING_RIGHT = 175,
-		TYPE_SIDEWALK_PERPENDICULAR_HALF_PARKING_LEFT = 176,
-		TYPE_SIDEWALK_PERPENDICULAR_HALF_PARKING_RIGHT = 177,
-		TYPE_SIDEWALK_PERPENDICULAR_PARKING_LEFT = 178,
-		TYPE_SIDEWALK_PERPENDICULAR_PARKING_RIGHT = 179,
-		TYPE_LIVING_STREET_BEGIN = 77,
-		TYPE_LIVING_STREET_END = 78,
-		TYPE_TUNNEL = 79,
-		TYPE_EMERGENCY_STOPPING_LEFT = 80,
-		TYPE_EMERGENCY_STOPPING_RIGHT = 81,
-		TYPE_HIGHWAY_BEGIN = 82,
-		TYPE_HIGHWAY_END = 83,
-		TYPE_EXPRESSWAY_BEGIN = 84,
-		TYPE_EXPRESSWAY_END = 85,
-		TYPE_NAMED_HIGHWAY_EXIT = 183,
-		TYPE_NAMED_EXPRESSWAY_EXIT = 184,
-		TYPE_NAMED_ROAD_EXIT = 185,
-		TYPE_HIGHWAY_EXIT = 86,
-		TYPE_EXPRESSWAY_EXIT = 186,
-		TYPE_ONEWAY_STREET = 187,
-		TYPE_CROSSING_GUARDS = 189,
-		TYPE_DEADEND = 190,
-		TYPE_DEADEND_EXCLUDING_DESIGNATED_ACTORS = 191,
-		TYPE_FIRST_AID_STATION = 194,
-		TYPE_POLICE_STATION = 195,
-		TYPE_TELEPHONE = 196,
-		TYPE_FILLING_STATION = 198,
-		TYPE_HOTEL = 201,
-		TYPE_INN = 202,
-		TYPE_KIOSK = 203,
-		TYPE_TOILET = 204,
-		TYPE_CHAPEL = 205,
-		TYPE_TOURIST_INFO = 206,
-		TYPE_REPAIR_SERVICE = 207,
-		TYPE_PEDESTRIAN_UNDERPASS = 208,
-		TYPE_PEDESTRIAN_BRIDGE = 209,
-		TYPE_CAMPER_PLACE = 213,
-		TYPE_ADVISORY_SPEED_LIMIT_BEGIN = 214,
-		TYPE_ADVISORY_SPEED_LIMIT_END = 215,
-		TYPE_PLACE_NAME = 216,
-		TYPE_TOURIST_ATTRACTION = 217,
-		TYPE_TOURIST_ROUTE = 218,
-		TYPE_TOURIST_AREA = 219,
-		TYPE_SHOULDER_NOT_PASSABLE_MOTOR_VEHICLES = 220,
-		TYPE_SHOULDER_UNSAFE_TRUCKS_TRACTORS = 221,
-		TYPE_TOLL_BEGIN = 222,
-		TYPE_TOLL_END = 223,
-		TYPE_TOLL_ROAD = 224,
-		TYPE_CUSTOMS = 225,
-		TYPE_INTERNATIONAL_BORDER_INFO = 226,
-		TYPE_STREETLIGHT_RED_BAND = 227,
-		TYPE_FEDERAL_HIGHWAY_ROUTE_NUMBER = 228,
-		TYPE_HIGHWAY_ROUTE_NUMBER = 229,
-		TYPE_HIGHWAY_INTERCHANGE_NUMBER = 230,
-		TYPE_EUROPEAN_ROUTE_NUMBER = 231,
-		TYPE_FEDERAL_HIGHWAY_DIRECTION_LEFT = 232,
-		TYPE_FEDERAL_HIGHWAY_DIRECTION_RIGHT = 233,
-		TYPE_PRIMARY_ROAD_DIRECTION_LEFT = 234,
-		TYPE_PRIMARY_ROAD_DIRECTION_RIGHT = 235,
-		TYPE_SECONDARY_ROAD_DIRECTION_LEFT = 236,
-		TYPE_SECONDARY_ROAD_DIRECTION_RIGHT = 237,
-		TYPE_DIRECTION_DESIGNATED_ACTORS_LEFT = 238,
-		TYPE_DIRECTION_DESIGNATED_ACTORS_RIGHT = 239,
-		TYPE_ROUTING_DESIGNATED_ACTORS = 240,
-		TYPE_DIRECTION_TO_HIGHWAY_LEFT = 143,
-		TYPE_DIRECTION_TO_HIGHWAY_RIGHT = 108,
-		TYPE_DIRECTION_TO_LOCAL_DESTINATION_LEFT = 127,
-		TYPE_DIRECTION_TO_LOCAL_DESTINATION_RIGHT = 136,
-		TYPE_CONSOLIDATED_DIRECTIONS = 118,
-		TYPE_STREET_NAME = 119,
-		TYPE_DIRECTION_PREANNOUNCEMENT = 120,
-		TYPE_DIRECTION_PREANNOUNCEMENT_LANE_CONFIG = 121,
-		TYPE_DIRECTION_PREANNOUNCEMENT_HIGHWAY_ENTRIES = 122,
-		TYPE_HIGHWAY_ANNOUNCEMENT = 123,
-		TYPE_OTHER_ROAD_ANNOUNCEMENT = 124,
-		TYPE_HIGHWAY_ANNOUNCEMENT_TRUCK_STOP = 125,
-		TYPE_HIGHWAY_PREANNOUNCEMENT_DIRECTIONS = 126,
-		TYPE_POLE_EXIT = 88,
-		TYPE_HIGHWAY_DISTANCE_BOARD = 180,
-		TYPE_DETOUR_LEFT = 181,
-		TYPE_DETOUR_RIGHT = 182,
-		TYPE_NUMBERED_DETOUR = 131,
-		TYPE_DETOUR_BEGIN = 132,
-		TYPE_DETOUR_END = 133,
-		TYPE_DETOUR_ROUTING_BOARD = 134,
-		TYPE_OPTIONAL_DETOUR = 111,
-		TYPE_OPTIONAL_DETOUR_ROUTING = 199,
-		TYPE_ROUTE_RECOMMENDATION = 211,
-		TYPE_ROUTE_RECOMMENDATION_END = 212,
-		TYPE_ANNOUNCE_LANE_TRANSITION_LEFT = 192,
-		TYPE_ANNOUNCE_LANE_TRANSITION_RIGHT = 193,
-		TYPE_ANNOUNCE_RIGHT_LANE_END = 90,
-		TYPE_ANNOUNCE_LEFT_LANE_END = 89,
-		TYPE_ANNOUNCE_RIGHT_LANE_BEGIN = 115,
-		TYPE_ANNOUNCE_LEFT_LANE_BEGIN = 116,
-		TYPE_ANNOUNCE_LANE_CONSOLIDATION = 117,
-		TYPE_DETOUR_CITY_BLOCK = 142,
-		TYPE_GATE = 141,
-		TYPE_POLE_WARNING = 91,
-		TYPE_TRAFFIC_CONE = 140,
-		TYPE_MOBILE_LANE_CLOSURE = 139,
-		TYPE_REFLECTOR_POST = 114,
-		TYPE_DIRECTIONAL_BOARD_WARNING = 113,
-		TYPE_GUIDING_PLATE = 104,
-		TYPE_GUIDING_PLATE_WEDGES = 105,
-		TYPE_PARKING_HAZARD = 99,
-		TYPE_TRAFFIC_LIGHT_GREEN_ARROW = 92,
-		TrafficSign_MainSign_Classification_Type_INT_MIN_SENTINEL_DO_NOT_USE_
-		= std::numeric_limits<int32_t>::min(),
-		TrafficSign_MainSign_Classification_Type_INT_MAX_SENTINEL_DO_NOT_USE_
-		= std::numeric_limits<int32_t>::max()
-	};
-
-	Signal(double s,
-		   double t,
-		   int id,
-		   std::string name,
-		   bool dynamic,
-		   Orientation orientation,
-		   double z_offset,
-		   std::string country,
-		   int type,
-		   double value,
-		   std::string unit,
-		   double height,
-		   double width,
-		   std::string text,
-		   double h_offset,
-		   double pitch,
-		   double roll)
-		: s_(s),
-		  t_(t),
-		  id_(id),
-		  name_(name),
-		  dynamic_(dynamic),
-		  orientation_(orientation),
-		  z_offset_(z_offset),
-		  country_(country),
-		  type_(type),
-		  value_(value),
-		  unit_(unit),
-		  height_(height),
-		  width_(width),
-		  text_(text),
-		  h_offset_(h_offset),
-		  pitch_(pitch),
-		  roll_(roll),
-		  length_(0) {}
-
-	std::string GetName() { return name_; }
-	int GetId() { return id_; }
-	double GetS() { return s_; }
-	double GetT() { return t_; }
-	void SetLength(double length) { length_ = length; }
-	double GetLength() { return length_; }
-	double GetHOffset() { return h_offset_; }
-	double GetZOffset() { return z_offset_; }
-	Orientation GetOrientation() { return orientation_; }
-	int GetType() { return type_; }
-	double GetHeight() { return height_; }
-	double GetWidth() { return width_; }
-	bool IsDynamic() { return dynamic_; }
-	double GetPitch() { return pitch_; }
-	double GetRoll() { return roll_; }
-	double GetValue() { return value_; }
-	std::string GetUnit() { return unit_; }
-	std::string GetText() { return text_; }
-	static Type GetTypeFromString(const std::string& type);
-	void Save(pugi::xml_node& signals);
-
-   private:
-	double s_;
-	double t_;
-	int id_;
-	std::string name_;
-	bool dynamic_;
-	Orientation orientation_;
-	double z_offset_;
-	std::string country_;
-	int type_;
-	double value_;
-	std::string unit_;
-	double height_;
-	double width_;
-	std::string text_;
-	double h_offset_;
-	double pitch_;
-	double roll_;
-	double length_;
-	static const std::map<std::string, Type> types_mapping_;
-};
-
-class OutlineCorner {
-   public:
-	virtual void GetPos(double& x, double& y, double& z) = 0;
-	virtual double GetHeight() = 0;
-	virtual ~OutlineCorner() {}
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	virtual void Save(pugi::xml_node&){};
-
-   private:
-	std::vector<UserData*> user_data_;
-};
-
-class OutlineCornerRoad : public OutlineCorner {
-   public:
-	OutlineCornerRoad(int roadId, double s, double t, double dz, double height);
-	OutlineCornerRoad(int id, int roadId, double s, double t, double dz, double height);
-	void GetPos(double& x, double& y, double& z);
-	double GetHeight() { return height_; }
-	void Save(pugi::xml_node&) override;
-
-   private:
-	int id_, roadId_;
-	double s_, t_, dz_, height_;
-};
-
-class OutlineCornerLocal : public OutlineCorner {
-   public:
-	OutlineCornerLocal(int roadId,
-					   double s,
-					   double t,
-					   double u,
-					   double v,
-					   double zLocal,
-					   double height,
-					   double heading);
-	OutlineCornerLocal(int id,
-					   int roadId,
-					   double s,
-					   double t,
-					   double u,
-					   double v,
-					   double zLocal,
-					   double height,
-					   double heading);
-	void GetPos(double& x, double& y, double& z);
-	double GetHeight() { return height_; }
-	void Save(pugi::xml_node&) override;
-
-   private:
-	int id_, roadId_;
-	double s_, t_, u_, v_, zLocal_, height_, heading_;
-};
-
-class Outline {
-   public:
-	typedef enum {
-		FILL_TYPE_GRASS,
-		FILL_TYPE_CONCRETE,
-		FILL_TYPE_COBBLE,
-		FILL_TYPE_ASPHALT,
-		FILL_TYPE_PAVEMENT,
-		FILL_TYPE_GRAVEL,
-		FILL_TYPE_SOIL,
-		FILL_TYPE_UNDEFINED
-	} FillType;
-
-	int id_;
-	FillType fillType_;
-	bool closed_;
-	std::vector<OutlineCorner*> corner_;
-	std::vector<UserData*> user_data_;
-
-	Outline(int id, FillType fillType, bool closed) : id_(id), fillType_(fillType), closed_(closed) {}
-
-	~Outline() {
-		for (size_t i = 0; i < corner_.size(); i++)
-			delete (corner_[i]);
-		corner_.clear();
-	}
-
-	void AddCorner(OutlineCorner* outlineCorner) { corner_.push_back(outlineCorner); }
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node&);
-};
-
-class Repeat {
-   public:
-	double s_;
-	double length_;
-	double distance_;
-	double tStart_;
-	double tEnd_;
-	double heightStart_;
-	double heightEnd_;
-	double zOffsetStart_;
-	double zOffsetEnd_;
-	double widthStart_;
-	double widthEnd_;
-	double lengthStart_;
-	double lengthEnd_;
-	double radiusStart_;
-	double radiusEnd_;
-
-	Repeat(double s,
-		   double length,
-		   double distance,
-		   double tStart,
-		   double tEnd,
-		   double heightStart,
-		   double heightEnd,
-		   double zOffsetStart,
-		   double zOffsetEnd)
-		: s_(s),
-		  length_(length),
-		  distance_(distance),
-		  tStart_(tStart),
-		  tEnd_(tEnd),
-		  heightStart_(heightStart),
-		  heightEnd_(heightEnd),
-		  zOffsetStart_(zOffsetStart),
-		  zOffsetEnd_(zOffsetEnd),
-		  widthStart_(0.0),
-		  widthEnd_(0.0),
-		  lengthStart_(0.0),
-		  lengthEnd_(0.0),
-		  radiusStart_(0.0),
-		  radiusEnd_(0.0) {}
-
-	void SetWidthStart(double widthStart) { widthStart_ = widthStart; }
-	void SetWidthEnd(double widthEnd) { widthEnd_ = widthEnd; }
-	void SetLengthStart(double lengthStart) { lengthStart_ = lengthStart; }
-	void SetLengthEnd(double lengthEnd) { lengthStart_ = lengthEnd; }
-	void SetHeightStart(double heightStart) { heightStart_ = heightStart; }
-	void SeHeightEnd(double heightStart) { heightStart_ = heightStart; }
-	double GetS() { return s_; }
-	double GetLength() { return length_; }
-	double GetDistance() { return distance_; }
-	double GetTStart() { return tStart_; }
-	double GetTEnd() { return tEnd_; }
-	double GetHeightStart() { return heightStart_; }
-	double GetHeightEnd() { return heightEnd_; }
-	double GetZOffsetStart() { return zOffsetStart_; }
-	double GetZOffsetEnd() { return zOffsetEnd_; }
-	double GetWidthStart() { return widthStart_; }
-	double GetWidthEnd() { return widthEnd_; }
-	double GetLengthStart() { return lengthStart_; }
-	double GetLengthEnd() { return lengthEnd_; }
-	double GetRadiusStart() { return radiusStart_; }
-	double GetRadiusEnd() { return radiusEnd_; }
-
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node&);
-
-   private:
-	std::vector<UserData*> user_data_;
-};
-
-class RMObject : public RoadObject {
-   public:
-	RMObject(double s,
-			 double t,
-			 int id,
-			 std::string name,
-			 Orientation orientation,
-			 double z_offset,
-			 std::string type,
-			 double length,
-			 double height,
-			 double width,
-			 double heading,
-			 double pitch,
-			 double roll)
-		: s_(s),
-		  t_(t),
-		  id_(id),
-		  name_(name),
-		  orientation_(orientation),
-		  z_offset_(z_offset),
-		  type_(type),
-		  length_(length),
-		  height_(height),
-		  width_(width),
-		  heading_(heading),
-		  pitch_(pitch),
-		  roll_(roll),
-		  repeat_(0) {}
-
-	~RMObject() {
-		for (size_t i = 0; i < outlines_.size(); i++)
-			delete (outlines_[i]);
-		outlines_.clear();
-	}
-
-	std::string GetName() { return name_; }
-	std::string GetType() { return type_; }
-	int GetId() { return id_; }
-	double GetS() { return s_; }
-	double GetT() { return t_; }
-	double GetHOffset() { return heading_; }
-	double GetPitch() { return pitch_; }
-	double GetRoll() { return roll_; }
-	double GetZOffset() { return z_offset_; }
-	double GetHeight() { return height_; }
-	double GetLength() { return length_; }
-	double GetWidth() { return width_; }
-	Orientation GetOrientation() { return orientation_; }
-	void AddOutline(Outline* outline) { outlines_.push_back(outline); }
-	void SetRepeat(Repeat* repeat);								   // odr1.5
-	void AddRepeat(Repeat* repeat) { repeat_.push_back(repeat); }  // odr1.4
-	Repeat* GetRepeat();
-	int GetNumberOfOutlines() { return (int)outlines_.size(); }
-	Outline* GetOutline(int i) { return (0 <= i && i < outlines_.size()) ? outlines_[i] : 0; }
-	void Save(pugi::xml_node&);
-
-   private:
-	std::string type_;
-	std::string name_;
-	int id_;
-	double s_;
-	double t_;
-	double z_offset_;
-	Orientation orientation_;
-	double length_;
-	double height_;
-	double width_;
-	double heading_;
-	double pitch_;
-	double roll_;
-	std::vector<Outline*> outlines_;
-	std::vector<Repeat*> repeat_;  // OpenDRIVE 1.4 uses multiple repeat tags.
-};
-
-class LaneRoadLaneConnection {
-   public:
-	LaneRoadLaneConnection()
-		: lane_id_(0),
-		  connecting_road_id_(-1),
-		  connecting_lane_id_(0),
-		  contact_point_(ContactPointType::CONTACT_POINT_UNDEFINED) {}
-	LaneRoadLaneConnection(int lane_id, int connecting_road_id, int connecting_lane_id)
-		: lane_id_(lane_id),
-		  connecting_road_id_(connecting_road_id),
-		  connecting_lane_id_(connecting_lane_id),
-		  contact_point_(ContactPointType::CONTACT_POINT_UNDEFINED) {}
-	void SetLane(int id) { lane_id_ = id; }
-	void SetConnectingRoad(int id) { connecting_road_id_ = id; }
-	void SetConnectingLane(int id) { connecting_lane_id_ = id; }
-	int GetLaneId() { return lane_id_; }
-	int GetConnectingRoadId() { return connecting_road_id_; }
-	int GetConnectinglaneId() { return connecting_lane_id_; }
-
-	ContactPointType contact_point_;
-
-   private:
-	int lane_id_;
-	int connecting_road_id_;
-	int connecting_lane_id_;
-};
-
-class JunctionLaneLink {
-   public:
-	JunctionLaneLink(int from, int to) : from_(from), to_(to) {}
-	int from_;
-	int to_;
-	void Print() { printf("JunctionLaneLink: from %d to %d\n", from_, to_); }
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node& connection);
-
-   private:
-	std::vector<UserData*> user_data_;
-};
-
-class Connection {
-   public:
-	Connection(Road* incoming_road, Road* connecting_road, ContactPointType contact_point);
-	Connection(int id, Road* incoming_road, Road* connecting_road, ContactPointType contact_point);
-	~Connection();
-	int GetNumberOfLaneLinks() { return (int)lane_link_.size(); }
-	JunctionLaneLink* GetLaneLink(int idx) { return lane_link_[idx]; }
-	int GetConnectingLaneId(int incoming_lane_id);
-	Road* GetIncomingRoad() { return incoming_road_; }
-	Road* GetConnectingRoad() { return connecting_road_; }
-	ContactPointType GetContactPoint() { return contact_point_; }
-	void AddJunctionLaneLink(int from, int to);
-	void Print();
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node& junction);
-
-   private:
-	Road* incoming_road_;
-	Road* connecting_road_;
-	ContactPointType contact_point_;
-	std::vector<JunctionLaneLink*> lane_link_;
-	int id_;
-	std::vector<UserData*> user_data_;
-};
-
-typedef struct {
-	std::vector<UserData*> user_data_;
-	int signalId_;
-	std::string type_;
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node&);
-} Control;
-
-class Controller {
-   public:
-	Controller() : id_(0), name_(""), sequence_(0) {}
-	Controller(int id, std::string name, int sequence = 0) : id_(id), name_(name), sequence_(sequence) {}
-
-	void AddControl(Control ctrl) { control_.push_back(ctrl); }
-	int GetNumberOfControls() { return (int)control_.size(); }
-	Control* GetControl(int index) {
-		return ((index >= 0 && index < control_.size()) ? &control_[index] : nullptr);
-	}
-
-	int GetId() { return id_; }
-	std::string GetName() { return name_; }
-	int GetSequence() { return sequence_; }
-
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node& root);
-
-   private:
-	int id_;
-	std::string name_;
-	int sequence_;
-	std::vector<Control> control_;
-	std::vector<UserData*> user_data_;
-};
-
-typedef struct {
-	std::vector<UserData*> user_data_;
-	int id_;
-	std::string type_;
-	int sequence_;
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node& junction);
-} JunctionController;
-
-class Junction {
-   public:
-	typedef enum {
-		DEFAULT,
-		DIRECT,
-		VIRTUAL	 // not supported yet
-	} JunctionType;
-
-	typedef enum {
-		RANDOM,
-		SELECTOR_ANGLE,	 // choose road which heading (relative incoming road) is closest to specified angle
-	} JunctionStrategyType;
-
-	Junction(int id, std::string name, JunctionType type) : id_(id), name_(name), type_(type) {
-		SetGlobalId();
-	}
-	~Junction();
-	int GetId() { return id_; }
-	std::string GetName() { return name_; }
-	int GetNumberOfConnections() { return (int)connection_.size(); }
-	int GetNumberOfRoadConnections(int roadId, int laneId);
-	LaneRoadLaneConnection GetRoadConnectionByIdx(int roadId,
-												  int laneId,
-												  int idx,
-												  int laneTypeMask = Lane::LaneType::LANE_TYPE_ANY_DRIVING);
-	void AddConnection(Connection* connection) { connection_.push_back(connection); }
-	int GetNoConnectionsFromRoadId(int incomingRoadId);
-	Connection* GetConnectionByIdx(int idx) { return connection_[idx]; }
-	int GetConnectingRoadIdFromIncomingRoadId(int incomingRoadId, int index);
-	void Print();
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node& root);
-	bool IsOsiIntersection();
-	int GetGlobalId() { return global_id_; }
-	void SetGlobalId();
-	int GetNumberOfControllers() { return (int)controller_.size(); }
-	JunctionController* GetJunctionControllerByIdx(int index);
-	void AddController(JunctionController controller) { controller_.push_back(controller); }
-	Road* GetRoadAtOtherEndOfConnectingRoad(Road* connecting_road, Road* incoming_road);
-	JunctionType GetType() { return type_; }
-
-   private:
-	std::vector<Connection*> connection_;
-	std::vector<JunctionController> controller_;
-	int id_;
-	int global_id_;
-	std::string name_;
-	JunctionType type_;
-	std::vector<UserData*> user_data_;
-};
-
-struct GeoReference {
-	double a_;
-	double axis_;
-	double b_;
-	std::string ellps_;
-	double k_;
-	double k_0_;
-	double lat_0_;
-	double lon_0_;
-	double lon_wrap_;
-	double over_;
-	std::string pm_;
-	std::string proj_;
-	std::string units_;
-	std::string vunits_;
-	double x_0_;
-	double y_0_;
-	std::string datum_;
-	std::string geo_id_grids_;
-	double zone_;
-	int towgs84_;
-	std::vector<UserData*> user_data_;
-
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node&) const;
-};
-
-struct OpenDriveOffset {
-	OpenDriveOffset() : x_(0), y_(0), z_(0), hdg_(0){};
-	double x_;
-	double y_;
-	double z_;
-	double hdg_;
-
-	std::vector<UserData*> user_data_;
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-	void Save(pugi::xml_node&) const;
-	bool isValid() const { return x_ || y_ || z_ || hdg_; };
-};
-
-struct OpenDriveHeader {
-	void Save(pugi::xml_node&) const;
-	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
-
-	uint revMajor_;
-	uint revMinor_;
-	std::string name_;
-	float version_;
-	std::string date_;
-	double north_;
-	double south_;
-	double east_;
-	double west_;
-	std::string vendor_;
-	OpenDriveOffset offset_;
-	GeoReference georeference_;
-	std::vector<UserData*> user_data_;
-};
-
-typedef struct {
-	double pos[3];		 // position, in global coordinate system
-	double heading;		 // road heading at steering target point
-	double pitch;		 // road pitch (inclination) at steering target point
-	double roll;		 // road roll (camber) at steering target point
-	double width;		 // lane width
-	double curvature;	 // road curvature at steering target point
-	double speed_limit;	 // speed limit given by OpenDRIVE type entry
-	int roadId;			 // road ID
-	int junctionId;		 // junction ID (-1 if not in a junction)
-	int laneId;			 // lane ID
-	double laneOffset;	 // lane offset (lateral distance from lane center)
-	double s;			 // s (longitudinal distance along reference line)
-	double t;			 // t (lateral distance from reference line)
-} RoadLaneInfo;
-
-typedef struct {
-	RoadLaneInfo road_lane_info;  // Road info at probe target position
-	double
-		relative_pos[3];  // probe target position relative vehicle (pivot position object) coordinate system
-	double relative_h;	  // heading angle to probe target from and relatove to vehicle (pivot position)
-} RoadProbeInfo;
-
-typedef struct {
-	double ds;		 // delta s (longitudinal distance)
-	double dt;		 // delta t (lateral distance)
-	int dLaneId;	 // delta laneId (increasing left and decreasing to the right)
-	double dx;		 // delta x (world coordinate system)
-	double dy;		 // delta y (world coordinate system)
-	double dxLocal;	 // delta x (local coordinate system)
-	double dyLocal;	 // delta y (local coordinate system)
-} PositionDiff;
-
-enum class CoordinateSystem { CS_UNDEFINED, CS_ENTITY, CS_LANE, CS_ROAD, CS_TRAJECTORY };
-
-enum class RelativeDistanceType {
-	REL_DIST_UNDEFINED,
-	REL_DIST_LATERAL,
-	REL_DIST_LONGITUDINAL,
-	REL_DIST_CARTESIAN,
-	REL_DIST_EUCLIDIAN
-};
-
-// Forward declarations
-class Route;
-class RMTrajectory;
-
-// A route is a sequence of positions, at least one per road along the route
-class Route {
-   public:
-	Route() : invalid_route_(false), waypoint_idx_(-1), path_s_(0), length_(0) {}
-
-	/**
-	Adds a waypoint to the route. One waypoint per road. At most one junction between waypoints.
-	@param position A regular position created with road, lane or world coordinates
-	@return Non zero return value indicates error of some kind
-	*/
-	int AddWaypoint(Position* position);
-	int GetWayPointDirection(int index);
-
-	void setName(std::string name);
-	std::string getName();
-	double GetLength() { return length_; }
-	void CheckValid();
-	bool IsValid() { return !invalid_route_; }
-
-	// Current route position data
-	// Actual object position might differ, e.g. laneId or even trackId in junctions
-	double GetTrackS() { return currentPos_.GetS(); }
-	double GetPathS() { return path_s_; }
-	int GetLaneId() { return currentPos_.GetLaneId(); }
-	int GetTrackId() { return currentPos_.GetTrackId(); }
-	Position* GetWaypoint(int index = -1);	// -1 means current
-	Road* GetRoadAtOtherEndOfConnectingRoad(Road* incoming_road);
-	int GetDirectionRelativeRoad();
-	Position* GetCurrentPosition() { return &currentPos_; }
-
-	/**
-	Specify route position in terms of a track ID and track S value
-	@return Non zero return value indicates error of some kind
-	*/
-	Position::ErrorCode SetTrackS(int trackId, double s);
-
-	/**
-	Move current position forward, or backwards, ds meters along the route
-	@param ds Distance to move, negative will move backwards
-	@param actualDistance Distance considering lateral offset and curvature (true/default) or along centerline
-	(false)
-	@return Non zero return value indicates error of some kind, most likely End Of Route
-	*/
-	Position::ErrorCode MovePathDS(double ds);
-
-	/**
-	Move current position to specified S-value along the route
-	@param route_s Distance to move, negative will move backwards
-	@return Non zero return value indicates error of some kind, most likely End Of Route
-	*/
-	Position::ErrorCode SetPathS(double s);
-
-	Position::ErrorCode CopySFractionOfLength(Position* pos);
-
-	std::vector<Position> minimal_waypoints_;  // used only for the default controllers
-	std::vector<Position> all_waypoints_;	   // used for user-defined controllers
-	std::string name_;
-	bool invalid_route_;
-	double path_s_;
-	Position currentPos_;
-	double length_;
-	int waypoint_idx_;
-};
-
-// A Road Path is a linked list of road links (road connections or junctions)
-// between a starting position and a target position
-// The path can be calculated automatically
-class RoadPath {
-   public:
-	typedef struct PathNode {
-		RoadLink* link;
-		double dist;
-		Road* fromRoad;
-		int fromLaneId;
-		ContactPointType contactPoint;
-		PathNode* previous;
-	} PathNode;
-
-	std::vector<PathNode*> visited_;
-	std::vector<PathNode*> unvisited_;
-	const Position* startPos_;
-	const Position* targetPos_;
-	int direction_;	 // direction of path from starting pos. 0==not set, 1==forward, 2==backward
-
-	RoadPath(const Position* startPos, const Position* targetPos)
-		: startPos_(startPos), targetPos_(targetPos){};
-	~RoadPath();
-
-	/**
-	Calculate shortest path between starting position and target position,
-	using Dijkstra's algorithm https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
-	it also calculates the length of the path, or distance between the positions
-	positive distance means that the shortest path was found in forward direction
-	negative distance means that the shortest path goes in opposite direction from the heading of the starting
-	position
-	@param dist A reference parameter into which the calculated path distance is stored
-	@param bothDirections Set to true in order to search also backwards from object
-	@param maxDist If set the search along each path branch will terminate after reaching this distance
-	@return 0 on success, -1 on failure e.g. path not found
-	*/
-	int Calculate(double& dist, bool bothDirections = true, double maxDist = LARGE_NUMBER);
-
-   private:
-	bool CheckRoad(Road* checkRoad, RoadPath::PathNode* srcNode, Road* fromRoad, int fromLaneId);
-};
-
-typedef struct {
-	double s;
-	double x;
-	double y;
-	double z;
-	double h;
-	double time;
-	double speed;
-	double p;
-	bool calcHeading;
-} TrajVertex;
-
-class PolyLineBase {
-   public:
-	PolyLineBase()
-		: length_(0), vIndex_(0), currentPos_({0, 0, 0, 0, 0, 0, 0, 0, false}), interpolateHeading_(false) {}
-	TrajVertex* AddVertex(TrajVertex p);
-	TrajVertex* AddVertex(double x, double y, double z, double h);
-	TrajVertex* AddVertex(double x, double y, double z);
-
-	/**
-	 * Update vertex position and recalculate dependent values, e.g. length and heading
-	 * NOTE: Need to be called in order, starting from i=0
-	 * @param i Index of vertex to update
-	 * @param x X coordinate of new position
-	 * @param y Y coordinate of new position
-	 * @param z Z coordinate of new position
-	 * @param h Heading
-	 */
-	TrajVertex* UpdateVertex(int i, double x, double y, double z, double h);
-
-	/**
-	 * Update vertex position and recalculate dependent values, e.g. length and heading
-	 * NOTE: Need to be called in order, starting from i=0
-	 * @param i Index of vertex to update
-	 * @param x X coordinate of new position
-	 * @param y Y coordinate of new position
-	 * @param z Z coordinate of new position
-	 */
-	TrajVertex* UpdateVertex(int i, double x, double y, double z);
-
-	void reset() { length_ = 0.0; }
-	int Evaluate(double s, TrajVertex& pos, double cornerRadius, int startAtIndex);
-	int Evaluate(double s, TrajVertex& pos, double cornerRadius);
-	int Evaluate(double s, TrajVertex& pos, int startAtIndex);
-	int Evaluate(double s, TrajVertex& pos);
-	int FindClosestPoint(double xin, double yin, TrajVertex& pos, int& index, int startAtIndex = 0);
-	int FindPointAhead(double s_start, double distance, TrajVertex& pos, int& index, int startAtIndex = 0);
-	int GetNumberOfVertices() { return (int)vertex_.size(); }
-	TrajVertex* GetVertex(int index);
-	void Reset();
-	int Time2S(double time, double& s);
-
-	std::vector<TrajVertex> vertex_;
-	TrajVertex currentPos_;
-	double length_;
-	int vIndex_;
-	bool interpolateHeading_;
-
-   protected:
-	int EvaluateSegmentByLocalS(int i, double local_s, double cornerRadius, TrajVertex& pos);
-};
-
-// Trajectory stuff
-class Shape {
-   public:
-	typedef enum { POLYLINE, CLOTHOID, NURBS, SHAPE_TYPE_UNDEFINED } ShapeType;
-
-	typedef enum { TRAJ_PARAM_TYPE_S, TRAJ_PARAM_TYPE_TIME } TrajectoryParamType;
-
-	Shape(ShapeType type) : type_(type) {}
-	virtual int Evaluate(double p, TrajectoryParamType ptype, TrajVertex& pos) { return -1; };
-	int FindClosestPoint(double xin, double yin, TrajVertex& pos, int& index, int startAtIndex = 0) {
-		return -1;
-	};
-	virtual double GetLength() { return 0.0; }
-	virtual double GetStartTime() { return 0.0; }
-	virtual double GetDuration() { return 0.0; }
-	ShapeType type_;
-
-	PolyLineBase pline_;  // approximation of shape, used for calculations and visualization
-};
-
-class PolyLineShape : public Shape {
-   public:
-	class Vertex {
-	   public:
-		Position pos_;
-	};
-
-	PolyLineShape() : Shape(ShapeType::POLYLINE) {}
-	void AddVertex(Position pos, double time, bool calculateHeading);
-	int Evaluate(double p, TrajectoryParamType ptype, TrajVertex& pos);
-	double GetLength() { return pline_.length_; }
-	double GetStartTime();
-	double GetDuration();
-
-	std::vector<Vertex*> vertex_;
-};
-
-class ClothoidShape : public Shape {
-   public:
-	ClothoidShape(roadmanager::Position pos,
-				  double curv,
-				  double curvDot,
-				  double len,
-				  double tStart,
-				  double tEnd);
-
-	int Evaluate(double p, TrajectoryParamType ptype, TrajVertex& pos);
-	int EvaluateInternal(double s, TrajVertex& pos);
-	void CalculatePolyLine();
-	double GetLength() { return spiral_->GetLength(); }
-	double GetStartTime();
-	double GetDuration();
-
-	Position pos_;
-	roadmanager::Spiral* spiral_;  // make use of the OpenDRIVE clothoid definition
-	double t_start_;
-	double t_end_;
-};
-
-/**
-	This nurbs implementation is strongly inspired by the "Nurbs Curve Example" at:
-	https://nccastaff.bournemouth.ac.uk/jmacey/OldWeb/RobTheBloke/www/opengl_programming.html
-*/
-class NurbsShape : public Shape {
-	class ControlPoint {
-	   public:
-		Position pos_;
-		double time_;
-		double weight_;
-		double t_;
-		bool calcHeading_;
-
-		ControlPoint(Position pos, double time, double weight, bool calcHeading)
-			: pos_(pos), time_(time), weight_(weight), calcHeading_(calcHeading) {}
-	};
-
-   public:
-	NurbsShape(int order) : order_(order), Shape(ShapeType::NURBS), length_(0) {
-		pline_.interpolateHeading_ = true;
-	}
-
-	void AddControlPoint(Position pos, double time, double weight, bool calcHeading);
-	void AddKnots(std::vector<double> knots);
-	int Evaluate(double p, TrajectoryParamType ptype, TrajVertex& pos);
-	int EvaluateInternal(double s, TrajVertex& pos);
-
-	int order_;
-	std::vector<ControlPoint> ctrlPoint_;
-	std::vector<double> knot_;
-	std::vector<double> d_;	 // used for temporary storage of CoxDeBoor weigthed control points
-	std::vector<double>
-		dPeakT_;  // used for storage of at what t value the corresponding ctrlPoint contribution peaks
-	std::vector<double>
-		dPeakValue_;  // used for storage of at what t value the corresponding ctrlPoint contribution peaks
-
-	void CalculatePolyLine();
-	double GetLength() { return length_; }
-	double GetStartTime();
-	double GetDuration();
-
-   private:
-	double CoxDeBoor(double x, int i, int p, const std::vector<double>& t);
-	double length_;
-};
-
-class RMTrajectory {
-   public:
-	Shape* shape_;
-
-	RMTrajectory(Shape* shape, std::string name, bool closed) : shape_(shape), name_(name), closed_(closed) {}
-	RMTrajectory() : shape_(0), closed_(false) {}
-	void Freeze();
-	double GetLength() { return shape_ ? shape_->GetLength() : 0.0; }
-	double GetTimeAtS(double s);
-	double GetStartTime();
-	double GetDuration();
-
-	std::string name_;
-	bool closed_;
-};
-
-}  // namespace roadmanager
-
-#endif	// OPENDRIVE_HH_
