+++ EnvironmentSimulator/Modules/RoadManager/inc/OpenDrive.hpp
@@ -1,15 +1,15 @@
 #ifndef OPENDRIVE_HPP
 #define OPENDRIVE_HPP
-#include "RoadManager.hpp"
 #include <memory>
-// #include "Road.hpp"
-// #include "Junction.hpp"
-// #include "Controller.hpp"
-// #include "Userdata.hpp"
+#include <vector>
+#include "Controller.hpp"
+#include "Junction.hpp"
+#include "Road.hpp"
+#include "Userdata.hpp"
 /**
- * @brief  This class has as responsibility to read and parse opendrive files. this includes populate subclasses and give access to them 
- * Should be easy to inherit for other classes that want to add top level functionality or lower class level functionality. 
- * 
+ * @brief  This class has as responsibility to read and parse opendrive files(check that is accurate). this
+ * includes populate subclasses and give access to them Should be easy to inherit for other classes that want
+ * to add top level functionality or lower class level functionality.
  */
 class OpenDrive {
    public:
@@ -24,27 +24,19 @@ class OpenDrive {
 	*/
 	bool LoadOpenDriveFile(const char* filename, bool replace = true);
 
-	/**
-		Initialize the global ids 
-	*/ 
-	// void InitGlobalLaneIds(); should not be in this class
-
-	/**
-		Get the filename of currently loaded OpenDRIVE file
-	*/
 	std::string GetOpenDriveFilename() { return odr_filename_; }
 
 	/**
 		Setting information based on the OSI standards for OpenDrive elements
 	*/
 
-    // This is OSI stuff should be moved seperated move to osi class that inhereits odr class
-	// bool SetRoadOSI(); 
+	// This is OSI stuff should be moved seperated move to osi class that inhereits odr class
+	// bool SetRoadOSI();
 	// bool CheckLaneOSIRequirement(std::vector<double> x0,
-								//  std::vector<double> y0,
-								//  std::vector<double> x1,
-								//  std::vector<double> y1);
-    // void SetLaneOSIPoints();
+	//  std::vector<double> y0,
+	//  std::vector<double> x1,
+	//  std::vector<double> y1);
+	// void SetLaneOSIPoints();
 	// void SetRoadMarkOSIPoints();
 
 	/**
@@ -52,14 +44,14 @@ class OpenDrive {
 		then it creates a LaneBoundary following the lane border (left border for left lanes, right border for
 	   right lanes)
 	*/
-	// void SetLaneBoundaryPoints(); check for osi stuff move 
+	// void SetLaneBoundaryPoints(); check for osi stuff move
 
 	/**
 		Retrieve a road segment specified by road ID
 		@param id road ID as specified in the OpenDRIVE file
 	*/
 	Road* GetRoadById(int id);
-    
+
 	/**
 		Retrieve a road segment specified by road vector element index
 		useful for iterating over all available road segments, e.g:
@@ -70,23 +62,30 @@ class OpenDrive {
 		@param idx index into the vector of roads
 	*/
 	Road* GetRoadByIdx(int idx);
-	// if this is added this class will be huge becouse you will need to have getters for all subclass private variabels
-    // Geometry* GetGeometryByIdx(int road_idx, int geom_idx);
-    int GetRoadIdxById(int id);
-	int GetTrackIdByIdx(int idx);
+	// if this is added this class will be huge becouse you will need to have getters for all subclass private
+	// variabels Geometry* GetGeometryByIdx(int road_idx, int geom_idx);
+	int GetRoadIdxById(int id);
+	int GetRoadIdByIdx(int idx);
 	int GetNumOfRoads() { return (int)road_.size(); }
+	// Return the whole protected subtag vector instead of index and one object, if you don't want to
+	// inherit(recommend), and do multiple manipulations on multiple roads at the smae time.
+	std::vector < std::shared_ptr<Road>& getRoadVectorRef() { return road_; }
+	std::vector < std::shared_ptr<Junction>& getJunctionsVectorRef() { return junction_; }
+	std::vector < std::shared_ptr<Controller>& getControllerVectorRef() { return controller_; }
+	std::vector < std::shared_ptr<UserData>& getUserDataVectorRef() { return user_data_; }
+
 	Junction* GetJunctionById(int id);
 	Junction* GetJunctionByIdx(int idx);
 
 	int GetNumOfJunctions() { return (int)junction_.size(); }
 
-	bool IsIndirectlyConnected(int road1_id,
-							   int road2_id,
-							   int*& connecting_road_id,
-							   int*& connecting_lane_id,
-							   int lane1_id = 0,
-							   int lane2_id = 0);
-
+	//  Move to other class
+	// bool IsIndirectlyConnected(int road1_id,
+	//    int road2_id,
+	//    int*& connecting_road_id,
+	//    int*& connecting_lane_id,
+	//    int lane1_id = 0,
+	//    int lane2_id = 0);
 	/**
 		Add any missing connections so that road connectivity is two-ways
 		Look at all road connections, and make sure they are defined both ways
@@ -120,16 +119,23 @@ class OpenDrive {
 
 	OpenDriveHeader GetHeader() const { return header_; };
 	void AddUserData(UserData* userData) { user_data_.push_back(userData); }
+	/**
+		Initialize the global ids
+	*/
+	// void InitGlobalLaneIds(); should not be in this class
 
+	/**
+		Get the filename of currently loaded OpenDRIVE file
+	*/
    protected:
-    // implement share pointers we have , I am guessing we are leaking memory
-    // classes that inherits opendriveparser should have a easy way to access subclasses
-    std::vector<std::shared_ptr<Road>> road_;
-	std::vector<std::shared_ptr<Junction>> junction_;
-	std::vector<std::shared_ptr<Controler>> controller_;
-	std::vector<std::shared_ptr<UserData>> user_data_;
-	std::map<std::string, std::string> signals_types_;
-    GeoReference geo_ref_;	// TODO: Remove this and use the header container instead.
+	// implement share pointers we have , I am guessing we are leaking memory
+	// classes that inherits opendriveparser should have a easy way to access subclasses
+	std::vector<std::shared_ptr<Road>> road_;			   // TODO check if it can be unique ptr
+	std::vector<std::shared_ptr<Junction>> junction_;	   // TODO check if it can be unique ptr
+	std::vector<std::shared_ptr<Controller>> controller_;  // TODO check if it can be unique ptr
+	std::vector<std::shared_ptr<UserData>> user_data_;	   // TODO check if it can be unique ptr
+	std::map<std::string, std::string> signals_types_;	   // TODO check if it can be unique ptr
+	GeoReference geo_ref_;	// TODO: Remove this and use the header container instead.
 
    private:
 	pugi::xml_node root_node_;
