+++ EnvironmentSimulator/Modules/RoadManager/src/Geometry.cpp
