+++ EnvironmentSimulator/Modules/RoadManager/src/Polynomial.cpp
