+++ EnvironmentSimulator/Modules/RoadManager/src/odrSpiral.cpp
